moved angle cosine to sec_emission.py
